backend/requirements.txt
Django==4.1.5
psycopg2-binary==2.9.1
Pillow==8.3.2
django-admin-interface==0.20.0
djangorestframework==3.12.4
django-cors-headers==3.13.0
djangorestframework-simplejwt==4.8.0
django-filter==22.1
django-storages==1.13.1
Redis==4.2.0
django-redis==5.2.0
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
flower==1.2.0
django-environ==0.9.0
lxml==4.9.1
newspaper3k==0.2.8
goose3==3.1.12
langdetect==1.0.9
bs4
django_enum_choices
django-storages==1.13.1
htmldate
django-environ==0.9.0
python-magic==0.4.27
gunicorn==20.1.0
selenium==4.8.0